Project Initialization
This monolith commit comprises work from Friday the 23rd 2024 to Monday the 26th 2024. It involves a simple database setup, PHP MVC architecture with a request router, jQuery AJAX requests and some simple bootstrap styling and CSS. the Index file contains a list of the Todo Items (pun intended) for the project.